aula 1 - agentes de ia/ai-app-intro/pyproject.toml
[project]
name = "ai-app-intro"
version = "0.1.0"
description = ""
authors = [
    {name = "Ian Vieira Salgado",email = "vieira4ev@gmail.com"}
]
readme = "README.md"
requires-python = "~3.13"
dependencies = [
    "rich (>=14.2.0,<15.0.0)",
    "windows-curses (>=2.4.1,<3.0.0)",
    "openai (>=2.15.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "langchain (>=1.2.3,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-openai (>=1.1.7,<2.0.0)"
]

[tool.poetry]
packages = [{include = "ai_app_intro", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
